+++ VocaluxeLib/IBasic.cs
@@ -267,14 +267,14 @@ namespace VocaluxeLib
         int Load(string soundFile, bool loop = false, bool prescan = false);
         void SetPosition(int soundStream, float newPosition);
         void Play(int soundStream);
-        void Fade(int soundStream, float targetVolume, float duration, EStreamAction afterFadeAction = EStreamAction.Nothing);
+        void Fade(int soundStream, int targetVolume, float duration, EStreamAction afterFadeAction = EStreamAction.Nothing);
         void Close(int soundStream);
 
         bool IsFinished(int soundStream);
         float GetPosition(int soundStream);
         float GetLength(int soundStream);
 
-        void SetStreamVolume(int soundStream, float volume);
+        void SetStreamVolume(int soundStream, int volume);
         void SetGlobalVolume(int volume);
     }
